feat(hosp,japan): add Japan ICU data with national "critical" standard
Add national standard of critical (eitehr on a ventilator or in ICU) as
ICU data.

Many prefectures use a different standard because of patient in ICU may
be put into ICU for close monitoring if beds are vaccant and are not
critical at the moment. But the national standard indicates bed
utilization better.

But anyway, hospitalization data is already in a similar situation since
they put mild patients (even asymptomatic patients) with higher risk
into hospital when beds are vaccant or new infections are stable.